+++ internal/static/integrations/node_exporter/config.go
@@ -112,10 +112,8 @@ type Config struct {
 	FilesystemFSTypesExclude         string              `yaml:"filesystem_fs_types_exclude,omitempty"`
 	FilesystemMountPointsExclude     string              `yaml:"filesystem_mount_points_exclude,omitempty"`
 	FilesystemMountTimeout           time.Duration       `yaml:"filesystem_mount_timeout,omitempty"`
-	HWMonChipInclude                 string              `yaml:"hwmon_chip_include,omitempty"`
-	HWMonChipExclude                 string              `yaml:"hwmon_chip_exclude,omitempty"`
-	HWMonSensorInclude               string              `yaml:"hwmon_sensor_include,omitempty"`
-	HWMonSensorExclude               string              `yaml:"hwmon_sensor_exclude,omitempty"`
+	HwMonChipInclude                 string              `yaml:"hwmon_chip_include,omitempty"`
+	HwMonChipExclude                 string              `yaml:"hwmon_chip_exclude,omitempty"`
 	IPVSBackendLabels                []string            `yaml:"ipvs_backend_labels,omitempty"`
 	NTPIPTTL                         int                 `yaml:"ntp_ip_ttl,omitempty"`
 	NTPLocalOffsetTolerance          time.Duration       `yaml:"ntp_local_offset_tolerance,omitempty"`